README.md
The San Francisco Department of Public Health Flood Health Vulnerability Index is a composite index that measures the spatial distribution and relative vulnerability of San Francisco communities to the health impacts of flood inundation and extreme storms. The index is constructed using socioeconomic and demographic, exposure, health, and housing indicators and is intended to serve as a planning tool for health and climate adaptation.